requirements.txt
Django==4.2.24
djangorestframework==3.14.0
django-allauth==0.57.2
dj-database-url==2.1.0
boto3==1.34.51
supabase==1.2.0
python-pptx==1.0.2
Pillow>=10.4.0
celery==5.3.6
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-csp==3.7
django-environ==0.11.2
django-ratelimit==4.1.0
django-rosetta==0.9.9
django-storages==1.14.2
crispy-tailwind==0.5.0
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
redis==5.0.3
openai==2.13.0
google-genai==1.56.0
transformers==4.39.3
torch==2.2.2
sentence-transformers==2.6.1
python-docx==1.1.0
reportlab==4.1.0
weasyprint==61.2
pytesseract==0.3.10
PyPDF2==3.0.1
pypdfium2==5.2.0
xlsxwriter==3.2.9
filetype==1.2.0
langdetect==1.0.9
scikit-learn==1.8.0
faker==38.2.0
coverage==7.13.0
flake8==7.3.0
pytest==9.0.2
sentry-sdk==1.45.0